backend/pyproject.toml
[tool.poetry]
name = "nico-airbnb"
version = "0.1.0"
description = ""
authors = ["zelonKim <ksz18601@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
beautifulsoup4 = "^4.12.2"
pillow = "^10.1.0"
djangorestframework = "^3.14.0"
strawberry-graphql = {extras = ["debug-server"], version = "^0.213.0"}
strawberry-graphql-django = "^0.24.1"
pyjwt = "^2.8.0"
django-environ = "0.9.0"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
